requirements.txt
bcrypt==4.1.3
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cssselect==1.2.0
dnspython==2.6.1
Flask==3.0.3
Flask-Login==0.6.3
flask-sitemapper==1.8.0
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.2.2
Markdown==3.6
markdown2==2.4.13
MarkupSafe==2.1.5
packaging==24.1
pymongo==4.7.3
pyquery==2.0.0
python-dotenv==1.0.1
readtime==3.0.0
requests==2.32.3
soupsieve==2.5
urllib3==2.2.2
Werkzeug==3.0.3
